Create LICENSE.md
We are no lawyers, we are not coders (by main profession), we are biologists. We wish FlexiDot to be as unrestrictive as possible. And we don't know much about licenses. So, most of our libraries have BSD, one LGPL, so we are pushing LGPL, if we grasped the concept correctly.